the complexity cannot be solved generally.
specific things cannot be solved generally, but are part of a specific use case.
There seems to be no way how renderMethod combinators like "horizontal" or "vertical" can have the same renderparameters or srcInfo like "renderToBlock".
This means: there seems to be now way how horizontal and vertical can be chained, in a general way.
Therefore I tried this:
	to combine renderFunctions, one has to define many things manually. Therefore the combinators "horizontal" and "vertical" take some more (function) parameters.